refactor: restructure codebase to separate client and server logic, update package.json, split tests
